+++ styled_text/example/lib/main.dart
@@ -68,12 +68,45 @@ class DemoPage extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
+            // Simple formatted text
             StyledText(
               text: 'Test: <b>bold</b> text.',
               styles: {
                 'b': TextStyle(fontWeight: FontWeight.bold),
               },
             ),
+
+            // Text with quotes
+            const SizedBox(height: 20),
+            StyledText(
+              text: 'Quoted Test: <b>&quot;bold&quot;</b> text.',
+              styles: {
+                'b': TextStyle(fontWeight: FontWeight.bold),
+              },
+            ),
+
+            // Multiline text without breaks
+            const SizedBox(height: 20),
+            StyledText(
+              text: """Multiline text 
+(wo breaks)""",
+              styles: {
+                'b': TextStyle(fontWeight: FontWeight.bold),
+              },
+            ),
+
+            // Multiline text with breaks
+            const SizedBox(height: 20),
+            StyledText(
+              text: """Multiline text
+(with breaks)""",
+              isNewLineAsBreaks: true,
+              styles: {
+                'b': TextStyle(fontWeight: FontWeight.bold),
+              },
+            ),
+
+            // Custom tags styles
             const SizedBox(height: 20),
             StyledText(
               text: 'Test: <bold>bold</bold> and <red>red color</red> text.',
@@ -83,6 +116,8 @@ class DemoPage extends StatelessWidget {
                     TextStyle(fontWeight: FontWeight.bold, color: Colors.red),
               },
             ),
+
+            // Text with icon
             const SizedBox(height: 20),
             StyledText(
               text: 'Text with alarm <alarm/> icon.',
@@ -90,6 +125,8 @@ class DemoPage extends StatelessWidget {
                 'alarm': IconStyle(Icons.alarm),
               },
             ),
+
+            // Text with action
             const SizedBox(height: 20),
             StyledText(
               text: 'Text with <action>action</action> inside.',
@@ -100,6 +137,8 @@ class DemoPage extends StatelessWidget {
                 ),
               },
             ),
+
+            // Text with link
             const SizedBox(height: 20),
             StyledText(
               text: