+++ styled_text/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* The parameter isNewLineAsBreaks has been added, indicating not to ignore line breaks in the source text.
+* Added description of parameters via doc comments.
+
 ## 1.0.0+2
 
 * Allow XML special chars: &lt; (<), &amp; (&), &gt; (>), &quot; ("), and &apos; (')